CA-99 abstract out validator logic
