+++ include/csr_validator.hh
@@ -1,0 +1,20 @@
+#ifndef CSR_VALIDATOR
+#define CSR_VALIDATOR
+
+#include "helpers.hh"
+#include "camera_validator.hh"
+#include "texture_validator.hh"
+#include "material_validator.hh"
+#include "primitive_validator.hh"
+#include "instance_validator.hh"
+
+/**
+ * @brief Checks if a CSR file is correctly formatted
+ * 
+ * @param[in]   _file   The file being validated
+ * 
+ * @note exits if CSR file is invalid
+ */
+void isCSR(std::ifstream& _file);
+
+#endif
